pyproject.toml
[project]
name = "zugzwang"
readme = "README.md"
dynamic = ["version"]
authors = [
    { name = "Kilian Calefice" },
]
requires-python = ">=3.11"
dependencies = [
    "torch==2.10.0",
    "torchvision==0.25.0",
    "torchaudio==2.10.0",
    "python-chess==1.999",
    "uvicorn[standard]==0.34.0",
    "fastapi==0.119.0",
    "tqdm==4.67.3",
    "onnx==1.17.0",
    "onnxscript==0.6.2",
]

[project.optional-dependencies]
# install with `pip install ".[dev]"`
dev = [
  "black==25.9.0",
  "isort==7.0.0",
  "pylint==4.0.1",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu130"
name = "pytorch-cu130"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu130" }
torchvision = { index = "pytorch-cu130" }
torchaudio = { index = "pytorch-cu130" }

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]

[project.urls]
"Repository" = "https://github.com/creed-braton/zugzwang"
"Issues" = "https://github.com/creed-braton/zugzwang/issues"
